move skeletonization modules into skeletonization package
