+++ gorgon/gui/__init__.py
@@ -1,5 +1,5 @@
-from .binary_skeletonization import BinarySkeletonization
-from .grayscale_skeletonization import GrayscaleSkeletonization
+from .skeletonization.binary_skeletonization import BinarySkeletonization
+from .skeletonization.grayscale_skeletonization import GrayscaleSkeletonization
 from .sse_identification import SSEIdentification
 from .sse_correspondence import SSECorrespondence
 from .calpha_atom_placement import CalphaAtomPlacement
