+++ gorgon/gui/skeletonization/grayscale_skeletonization.py
@@ -3,10 +3,10 @@ import sys
 
 from PyQt4 import QtGui
 
-from .explorer.skeleton_viewer import SkeletonViewer
-from .skeletonization.grayscale_form import VolumeGrayscaleSkeletonizationForm
-from .explorer import Camera
-from .explorer.volume_viewer import VolumeViewer
+from ..explorer.skeleton_viewer import SkeletonViewer
+from ..skeletonization.grayscale_form import VolumeGrayscaleSkeletonizationForm
+from ..explorer import Camera
+from ..explorer.volume_viewer import VolumeViewer
 
 
 class GrayscaleSkeletonization(QtGui.QMainWindow):
